DB갈아 엎기 및 닉네임 관련 수정
