--- src/ProjectDBCode/MsgDAO.java
@@ -1,8 +1,0 @@
-package ProjectDBCode;
-
-import java.util.List;
-
-public interface MsgDAO {
-    void insertMsg(String msg, String mem_ip, String roomName);
-    List<String> getMsgList(String roomName);
-}
